README and data download changes

+++ scripts/download_data.sh
@@ -1,3 +1,5 @@
 cd ..
-rm -r ./data/FlowCasesDeidentify120519
+if [ -d "./data/FlowCasesDeidentify120519" ] 
+then
+    rm -r ./data/FlowCasesDeidentify120519
 gsutil -m cp -r gs://flowai-data/FlowCasesDeidentify120519 ./data/
